% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{font.families.google}
\alias{font.families.google}
\title{List Font Families Available in Google Fonts}
\usage{
font.families.google()
}
\value{
A character vector of available font family names in Google Fonts.
}
\description{
This function lists family names of the fonts that are currently
available in Google Fonts. When running this function for the first time, 
it may take a few seconds to fetch the font information database.
This function requires the \pkg{jsonlite} and \pkg{curl} packages.
}
\examples{
\dontrun{
font.families.google()
}

}
\seealso{
\code{\link{font.add.google}()}
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
