% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{font_families}
\alias{font_families}
\alias{font.families}
\title{List Font Families Loaded by 'sysfonts'}
\usage{
font_families()

font.families()
}
\value{
A character vector of available font family names.
}
\description{
The two versions of this function are equivalent, but the
"underscore" naming is preferred.

This function lists font families currently available that can be
used by \pkg{R2SWF} and \pkg{showtext} packages.
}
\details{
By default there are three font families loaded automatically,
i.e., "sans", "serif" and "mono". If one wants to use other fonts,
\code{\link{font_add}()} needs to be called
to register new fonts by specifying a family name and corresponding
font files. See \code{\link{font_add}()} for details about
the meaning of "family name" in this context, as well as
a complete example of registering and using a new font.
}
\examples{
font_families()

}
\seealso{
\code{\link{font_add}()}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
