% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{font_files}
\alias{font_files}
\alias{font.files}
\title{List Font Files Available in the Search Paths}
\usage{
font_files()

font.files()
}
\value{
A character vector of font filenames.
}
\description{
The two versions of this function are equivalent, but the
"underscore" naming is preferred.

This function lists font files in the search path that can be
loaded by \code{\link{font_add}()}.
Currently supported formats include TrueType fonts(*.ttf, *.ttc) and OpenType fonts(*.otf).
}
\examples{
font_files()

}
\seealso{
\code{\link{font_paths}()}, \code{\link{font_add}()}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
