% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{font_info_google}
\alias{font_info_google}
\title{Display Information of Available Google Fonts}
\usage{
font_info_google()
}
\value{
A data frame containing meta data of Google Fonts.
}
\description{
This function returns a data frame that contains the meta data of
font families available in Google Fonts, for example the family name,
available font face variants, the version number, etc. When running this
function for the first time, it may take a few seconds to fetch the
database. This function requires the \pkg{jsonlite} and \pkg{curl} packages.
}
\examples{
\dontrun{
font_info_google()
}

}
\seealso{
\code{\link{font_families_google}()}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
