% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{font.files}
\alias{font.files}
\title{List Font Files Available in the Search Paths}
\usage{
font.files()
}
\value{
A character vector of font filenames.
}
\description{
This function lists font files in the search path that can be
loaded by \code{\link{font.add}()}.
Currently supported formats include TrueType fonts(*.ttf, *.ttc) and OpenType fonts(*.otf).
}
\examples{
font.files()

}
\seealso{
\code{\link{font.paths}()}, \code{\link{font.add}()}
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}
