% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{font_families_google}
\alias{font_families_google}
\alias{font.families.google}
\title{List Font Families Available in Google Fonts}
\usage{
font_families_google()

font.families.google()
}
\value{
A character vector of available font family names in Google Fonts.
}
\description{
The two versions of this function are equivalent, but the
"underscore" naming is preferred.

This function lists family names of the fonts that are currently
available in Google Fonts. When running this function for the first time, 
it may take a few seconds to fetch the font information database.
This function requires the \pkg{jsonlite} and \pkg{curl} packages.
}
\examples{
\dontrun{
font_families_google()
}

}
\seealso{
\code{\link{font_add_google}()}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
